Client Hotel Fulfilment and Start Trip Shortening
Added Client helper class that stores allocation information for that
client. Includes things such as allocated start and end dates and
allocated hotel types. This is separate from their preferences.

The agent should now be aware of when it has all the required hotels
for an individual and let the airplane method know to buy the flights
for that client.

Made a start on shortening the trips when a bid fails. Though upon
actually starting to code, its actually a bit trickier than I had
initially thought. I’ve commented out the code for now so that it will
not run. The agent is in a functional state.