previousPrice init, hotel ownership factor
Init previous prices to 200 otherwise its normal 0 and we can end up
with an artificially high delta. Also factored in the cost of switching
trips and not using a hotel that is already allocated to you. Moving to
a free hotel is still free. This prevent the client trying to change
trips to frequently